added additional conditional logic

+++ client.js
@@ -2,23 +2,37 @@ const net = require('net');
 const convert = require('xml-js');
 const request = require('request');
 
+// **** FOR TESTING ENVIRONMENT ****
+
+//for observation
+let portPeekerIp = '75.151.123.46';
+let portPeekerPort = 5000;
+
+//testserver config variables
+let testPort = 9001;
+let testIp = 'localhost';
+
+//outside network
+let externalJumpUrl = 'http://75.151.123.42/live.json';
+
+// ****** FOR LIVE ENVIRONMENT *****
+
 // config vars for automation system TCP connection
 // wide orbit 
 let port = 11112;
 let ip = "10.150.54.202";
 
 // config vars for streaming audio device connection
-
 let streamingPort = 21117,
 		streamingIp = "10.150.55.108";
 
-// jumpgate configs
-let jumpUrl = 'https://10.150.55.212/live.json';
+		//inside network
+let internalJumpUrl = 'http://10.150.55.212/live.json';
 
 // create initial client for connecting to automation system and listening for XML
 const client = new net.Socket();
 
-client.connect(port, ip, function() {
+client.connect(testPort, testIp, function() {
 	console.log('Automation client connected.')
 	handleConnection(client);
 });
@@ -43,10 +57,12 @@ function handleConnection(conn) {
 		console.log("Data detected. Attempting to generate URL...");
 
 		try {
-      let stringifiedBuffer = d.toString('utf8');
+			console.log('stringifying things');
+			let stringifiedBuffer = d.toString('utf8');
 			
 			// helper that takes XML and converts to JSON
       function xmlToJson(xml) {
+				console.log('running xml to json');
         let convertedXml = convert.xml2json(stringifiedBuffer);
         return JSON.parse(convertedXml);
       };
@@ -56,35 +72,51 @@ function handleConnection(conn) {
           cart = xmlObj.elements[0].elements[7].elements[0].text,
           category = xmlObj.elements[0].elements[6].elements[0].text,
 					cat = xmlObj.elements[0].elements[12].elements[0].text.toLowerCase();
-					
-					// jumpgate
-					request(jumpUrl, function(error, body, response) {
-						try {
-							var parsedJson = JSON.parse(body);
-							var sec = parsedJson.isaSEC;
 
-							// url to be sent out to streaming audio device
-							let url = `http://www.97xonline.com?autoID=${cart}&autoCat=${category}&sec=` + sec + `&dur=${duration}&cat=${cat}`
-				
-							console.log(`URL being sent to streaming audio device: ${url}`);
-							
-							// send data to another device via tcp
-							const streamingClient = new net.Socket();
-							try {
-								streamingClient.connect(streamingPort, streamingIp, function() {
-									console.log("Attemping to send URL to streaming device..."); 
-									streamingClient.write(url);	
-									streamingClient.destroy();
-								});
+					var durationInSeconds = duration / 1000;
+					console.log('duration is: ', durationInSeconds);
+
+					if (durationInSeconds <= 10 && category == 'IM2') {
+						console.log('duration is less than or equal to 10 seconds');
+						console.log('requesting data');
+						request(externalJumpUrl, function(error, response, body) {
+							if (error) {
+								console.error(error);
 							}
-							catch(error) {
-								console.log(error);
-							};
-						} catch(e) {
-							console.error(e);
-						}
-						
-					});
+							try {
+								console.log('got the data,', typeof body);
+								var parse = JSON.parse(body);
+								console.log('parsed data successfully');
+								var sec = parse.isaSEC;
+
+								// url to be sent out to streaming audio device
+								let url = `http://www.97xonline.com?autoID=${cart}&autoCat=${category}&sec=` + sec + `&dur=${duration}&cat=${cat}`
+					
+								console.log(`URL being sent to streaming audio device: ${url}`);
+								
+								// send data to another device via tcp
+								const streamingClient = new net.Socket();
+								try {
+									console.log('connecting to streaming client');
+									// streamingClient.connect(streamingPort, streamingIp, function() {
+										streamingClient.connect(portPeekerPort, portPeekerIp, function() {
+										console.log("Attemping to send URL to streaming device..."); 
+										streamingClient.write(url);
+										console.log('successfully wrote to the client');
+										streamingClient.destroy();
+									});
+								}
+								catch(error) {
+									console.log(error);
+								};
+							} catch(e) {
+								console.error(e);
+							}
+							
+						});
+					} else {
+						console.log('duration is greater than 10 seconds');
+					}
 			
     }
 		catch(error) {
